Begin breaking code into up into a more logical structure
