+++ ensemble_tagger_implementation/tagset_maps.py
@@ -1,0 +1,55 @@
+swum_pos_dictionary = {
+    "VI":"V",
+    "NI":"N",
+    "PP":"P",
+    "CJ":"CJ",
+    "D":"D",
+    "DT":"DT",
+    "N":"N",
+    "NM":"NM",
+    "V":"V",
+    "VM":"VM",
+    "V3PS":"V",
+    "NPL":"NPL",
+    "PR":"PR",
+    "P":"P"
+}
+posse_pos_dictionary = {
+    "verb":"V",
+    "noun":"N",
+    "closedlist":"P",
+    "adjective":"NM",
+}
+stanford_pos_dictionary = {
+"CC":"CJ",
+"CD":"D",
+"DT":"DT",
+"EX":"N",
+"FW":"N",
+"IN":"P",
+"JJ":"NM",
+"JJR":"NM",
+"JJS":"NM",
+"LS":"N",
+"MD":"V",
+"NN":"N",
+"NNS":"NPL",
+"NNP":"N",
+"NNPS":"NPL",
+"PDT":"DT",
+"POS":"N",
+"PRP":"P",
+"PRP":"P",
+"RB":"VM",
+"RBR":"VM",
+"RBS":"VM",
+"RP":"N",
+"SYM":"N",
+"TO":"P",
+"UH":"N",
+"VB":"V",
+"WDT":"DT",
+"WP":"P",
+"WP,":"P",
+"WRB":"VM"
+}
